src/mcp_servers/youtube-mcp-server/pyproject.toml
[project]
name = "mcp-example"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "httpx>=0.28.1",
    "mcp[cli]>=1.8.1",
    "uvicorn>=0.34.2",
    "youtube-search>=2.1.2",
    "google-api-python-client>=2.118.0",
    "google-auth-oauthlib>=1.2.0",
    "youtube-transcript-api>=0.6.2",
]

[tool.setuptools]
py-modules = ["mcp_videos", "yt_helper", "video_api", "mcp_videos_api"]
